requirements.txt
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
openai==0.26.1
gunicorn==20.1.0